+++ clite_parser.ts
@@ -34,6 +34,15 @@ export type CliteRunConfig = {
    * import.meta to use : don't run if the file is imported, and use the basename of import.meta.url in the help
    */
   meta?: ImportMeta;
+  /**
+   * enable "--config <path>" to load json config before processing the args, Show in the help if it's a string
+   */
+  configCli?: boolean | string;
+
+  /**
+   * don't convert "true"/"false" to true/false in command arguments, and not to number after --
+   */
+  dontConvertCmdArgs?: boolean;
 };
 
 /**
@@ -52,6 +61,28 @@ export function cliteRun(obj: Obj, config?: CliteRunConfig): unknown {
         console.error(help);
         return help;
       } else {
+        if (
+          config?.configCli &&
+          getFieldNames(parseResult.options).includes("config")
+        ) {
+          // deno-lint-ignore no-explicit-any
+          if ((globalThis as any)["Deno"]?.args) {
+            const path = parseResult.options.config as string;
+            try {
+              const json = Deno.readTextFileSync(path);
+              const config = JSON.parse(json);
+              Object.assign(obj, config);
+              obj.config = undefined;
+            } catch (error) {
+              throw new Error(`Error while loading the config file "${path}"`, {
+                cause: { clite: true, error },
+              });
+            }
+          } else {
+            // TODO NodeJS implementation
+            throw new Error("Load config is not implemented on NodeJs");
+          }
+        }
         const command = parseResult.command ?? defaultMethod;
         if (!command) {
           throw new Error(`no method defined or no "main" method`, {
@@ -64,9 +95,13 @@ export function cliteRun(obj: Obj, config?: CliteRunConfig): unknown {
           });
         }
         fillFields(parseResult, obj);
-        return runCommand(obj, command, parseResult.commandArgs, config);
+        const commandArgs = config?.dontConvertCmdArgs
+          ? parseResult.commandArgs
+          : parseResult.commandArgs.map(convertCommandArg);
+        return runCommand(obj, command, commandArgs, config);
       }
-    } catch (e) {
+      // deno-lint-ignore no-explicit-any
+    } catch (e: any) {
       if (e.cause?.clite || config?.printHelpOnError) {
         console.error(bgRed(bold("An error occurred ! The help :")));
         console.error(help);
@@ -76,4 +111,19 @@ export function cliteRun(obj: Obj, config?: CliteRunConfig): unknown {
       throw e;
     }
   }
+}
+
+function convertCommandArg(v: string | number) {
+  if (typeof v === "string") {
+    if (v === "true") {
+      return true;
+    } else if (v === "false") {
+      return false;
+    } else {
+      if (!isNaN(v as unknown as number) && !isNaN(parseFloat(v))) {
+        return parseFloat(v);
+      }
+    }
+  }
+  return v;
 }