+++ examples/example-object.ts
@@ -13,3 +13,7 @@ cliteRun(new Tool());
 
 // ./example-object.ts --obj.foo.bar1 12 --obj.foo.bar2 45  --obj.toto 89 --obj.tata 77
 // main command Tool { obj: { foo: { bar1: 12, bar2: 45 }, toto: 89, tata: 77 } }
+
+/*
+
+*/