+++ examples/plain_object_lite.ts
@@ -6,7 +6,6 @@ cliteRun({
   main() {
     console.log("main command", this);
   },
-  _up_help: "create and start the services",
   up(svc: string, timeout = 10) {
     console.log("up command", { svc, timeout, retry: this.retry });
   },
@@ -14,3 +13,19 @@ cliteRun({
     console.log("down command", { svc, retry: this.retry });
   },
 });
+/*
+    $ ./plain_object_lite.ts --help
+    Usage: <Object file> [Options] [--] [command [command args]]
+
+    Commands:
+      main               [default]
+      up <svc> <timeout>
+      down <svc>
+
+    Options:
+     -h, --help  Show this help [default: false]
+         --retry                    [default: 2]
+
+    $ ./plain_object_lite.ts --retry 6 up foo 3
+    up command { svc: "foo", timeout: 3, retry: 6 }
+*/